Rephrase evil factions' victory condition
In my opinion, the evil factions' win condition was not phrased accurately. There is first the condition, for one of the evil factions, to have all other evil factions gone. Then they have to reach parity with the village (I believe 'reaching parity with' is a bit more clear than 'outnumering').