First version of Vampire and Vampire Master in 1 file
Following StormPooper's suggestion for the format, I combined the role descriptions for the original Vampire and the Vampire Master in one document. This makes the role description longer than most. One use of HTML I'm unsure about: the <br /> in the middle of the Notes <ul> environment (inside a <li> element). I simply found this on the internet and copied it.