First version of all Familiar types in one file
Following StormPooper's suggestion for the format, I combined the role descriptions for all Familiar types in one document. This makes the role description longer than most. The content of the document is based on the original Familiar's text, but with some slight changes in the writing style. On top of that, a lot of extra information is added to describe the Familiar's new types and other VM-related aspects.