Separated VM from Vamp pages: VM here
The combined page was considered too long and therefore not inviting to read. Thus the pages were separated. This page is the VM's: all content strictly relating to Vamp is removed. All mentioning of 'Vampire' is now changed to 'Vampire Master', to be very explicit in which of both roles is described (even if the described characteristic holds for all Vamp roles).