Vamp faction: elaborate on basics + include VM
The VM wasn't included in this overview yet, so I added them. I also went over the essential Vamp stuff a little bit more elaborately. I made a reference to the tips of the respective role pages, just as was done before, but now I did create a 'Tips & Considertaions' header for visibility...