Added 'Ability Type' section (moved info from Notes there), plus small change to tip
This change is based on StormPooper's review. I've moved an important part of classifying the Familiar by their ability type out of notes, and gave it its own section. Here is defined the abilities a Familiar can have, based on the Vampire type recruiting them, and the amount of Familiars a Vamp type can recruit is also stated. This significantly reduces the lenght of hte Notes section and hopefully makes it clearer this way. I also changed an OG Vamp tip: I stated before that Fam(Stalker) and Stalker were the same, but now I've emphsized that a Seer still knows the difference between these roles...