removed paragraph on similarity Vamp/VM
This change is based on StormPooper's review. He suggested I remove the paragraph that tries to clarify the distinction between Vampire and the Vampire Master (and then we have the faction 'Vampire(s)' as well, of course), because we can clarify this in the role sidebar rather than on this role description page.