New General section: Killing & Dying
This section aims to give a clear overview of the various ways that people can die in the game, and the types of kills involved in the process. Part of this same update will be to clearly describe in ALL roles and Items pages the type of kill it is, using an uniform language of the kill/death types that are defined here. Setting up this framework will hopefully allow new players to understand the Kill/Death mechanics more easily. In addition, new mechanics can more readily be understood by the community if they are defined in the same terminology.