Add root-relative paths for all `route: ` entries.
Something seems to have changed at some point, and without the leading `/`
some of the generated paths in link URLs are not being generated correctly
leading to the site attempting to link to incorrectly, such as if you are
on the items page you might get a link `/guides/items/online-werewolf-rules`
for the rules page.

A side effect of this, was that BloodmoonCult ended up being the site index
again, despite `1_general/1_basics.mdx` (Removing the `/`s seems to make
basics the root. Renaming these from `1_` to `aaa_` seems to have the
intended result of making this the first page, so I suspect some weird
sorting going on somewhere.

Also fixed the header linking to `/guides/guides/` not `/guides/`